+++ dash_uploader/__init__.py
@@ -5,7 +5,7 @@ import json
 
 import dash as _dash
 
-from dash_uploader.configure_upload import configure_upload
+from dash_uploader.configure_upload import configure_upload, configure_upload_flask
 from dash_uploader.callbacks import callback
 from dash_uploader.httprequesthandler import HttpRequestHandler
 from dash_uploader.upload import Upload
@@ -16,7 +16,7 @@ from ._build._imports_ import __all__ as build_all
 
 # Defines all exposed APIs of this package.
 __all__ = [
-    'configure_upload',
+    'configure_upload', 'configure_upload_flask',
     'callback',
     'HttpRequestHandler',
     'Upload'